requirements.txt
fastapi 
uvicorn 
rasa
google-cloud-speech
google-cloud-texttospeech
psycopg2-binary
sqlalchemy<2.0
pydantic 
requests
